+++ nuxeo-drive-client/nxdrive/engine/watcher/remote_watcher.py
@@ -3,6 +3,7 @@ import os
 import socket
 from datetime import datetime
 from httplib import BadStatusLine
+from logging import getLogger
 from time import sleep
 from urllib2 import HTTPError, URLError
 
@@ -15,10 +16,9 @@ from nxdrive.client.common import COLLECTION_SYNC_ROOT_FACTORY_NAME, \
 from nxdrive.client.remote_file_system_client import RemoteFileInfo
 from nxdrive.engine.activity import Action
 from nxdrive.engine.workers import EngineWorker, ThreadInterrupt
-from nxdrive.logging_config import get_logger
 from nxdrive.utils import current_milli_time, path_join
 
-log = get_logger(__name__)
+log = getLogger(__name__)
 
 
 class RemoteWatcher(EngineWorker):